fdsg

+++ src/main/java/ml/work/main/controllers/ArticuloManufacturadoController.java
@@ -1,10 +1,10 @@
 package ml.work.main.controllers;
 
 import java.util.ArrayList;
+import java.util.List;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,46 +15,62 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 
-import ml.work.main.entities.ArticuloManufacturado;
 import ml.work.main.dtos.ArticuloManufacturadoDTO;
 import ml.work.main.dtos.MensajeDTO;
+import ml.work.main.entities.ArticuloManufacturado;
 import ml.work.main.service.ArticuloManufacturadoService;
 
-@Controller
+
 @RestController
-//@CrossOrigin(origins = "*") 
-@RequestMapping(path = "api/v1/manufacturados")
+@CrossOrigin(origins = "*") 
+@RequestMapping(path = "/api/v1/manufacturados")
 public class ArticuloManufacturadoController implements ObjectController<ArticuloManufacturadoDTO>{
 
+	@Autowired
 	private ArticuloManufacturadoService articuloManufacturadoService;
 
-	public ArticuloManufacturadoController(ArticuloManufacturadoService articuloManufacturadoService) {
+	/*public ArticuloManufacturadoController(ArticuloManufacturadoService articuloManufacturadoService) {
 		this.articuloManufacturadoService = articuloManufacturadoService;
-	}
+	}*/
 	
 	@Override
-	//@CrossOrigin("*")
-	@GetMapping(path = "/")
-	public ArrayList<ArticuloManufacturadoDTO> getAll() {
-		return ResponseEntity.status(200).body(articuloManufacturadoService.getAll()).getBody();
-	}
+	@CrossOrigin("*")
+	@GetMapping(path = "/lista")
+	public ResponseEntity<List<ArticuloManufacturadoDTO>> getAll(){
+        List<ArticuloManufacturadoDTO> lista = articuloManufacturadoService.getAll();
+        return new ResponseEntity<List<ArticuloManufacturadoDTO>>(lista, HttpStatus.OK);
+    }
+	@CrossOrigin("*")
+	@GetMapping(path = "/listado")
+	public ResponseEntity<List<ArticuloManufacturado>> getLista(){
+        List<ArticuloManufacturado> lista = articuloManufacturadoService.obtenerTodos();
+        return new ResponseEntity<List<ArticuloManufacturado>>(lista, HttpStatus.OK);
+    }
+	
 	
 	@GetMapping(path = "/disponibles/{buscar}")
 	public ArrayList<ArticuloManufacturadoDTO> getEnMenu(@PathVariable boolean buscar) {
 		return ResponseEntity.status(200).body(articuloManufacturadoService.getEnMenu(buscar)).getBody();
 	}	
 	
-	@Override
-	@GetMapping(path = "/{id}")
-	public ArticuloManufacturadoDTO getOne(@PathVariable int id) {
+	
+	/*public ArticuloManufacturadoDTO getOne(@PathVariable int id) {
 		
 		return ResponseEntity.status(200).body(articuloManufacturadoService.getOne(id)).getBody();
+	}*/
+	@CrossOrigin("*")	
+	@GetMapping("/detalle/{id}")
+	public ResponseEntity<ArticuloManufacturadoDTO> getOne(@PathVariable int id){
+		if(!articuloManufacturadoService.existePorId(id))
+			return new ResponseEntity(new MensajeDTO("no existe ese articulo manufacturado"), HttpStatus.NOT_FOUND);
+	    ArticuloManufacturadoDTO manufacturado = articuloManufacturadoService.getOne(id);
+	    return new ResponseEntity<ArticuloManufacturadoDTO>(manufacturado, HttpStatus.OK);
 	}
 	
-	
-	@Override
+	/*@Override
 	@PostMapping("/")
 	@PreAuthorize("hasRole('ADMIN')")
 	public ResponseEntity save(@RequestBody ArticuloManufacturadoDTO body) {
@@ -65,22 +81,36 @@ public class ArticuloManufacturadoController implements ObjectController<Articul
 		} else {
 			return ResponseEntity.status(400).body("{'error' : 'bad request'}");
 		}
+	}*/
+	@Override
+	@PostMapping("nuevo")
+	//@PreAuthorize("hasRole('ADMIN')")
+	public ResponseEntity save(@RequestBody ArticuloManufacturadoDTO manuf){
+		 
+       if(StringUtils.isBlank(manuf.getNombre_articuloM()))
+            return new ResponseEntity(new MensajeDTO("el nombre es obligatorio"), HttpStatus.BAD_REQUEST);
+       if((Float)manuf.getPrecioM() == null || manuf.getPrecioM()==0)
+            return new ResponseEntity(new MensajeDTO("el precio es obligatorio"), HttpStatus.BAD_REQUEST);
+       if(articuloManufacturadoService.existePorNombre(manuf.getNombre_articuloM()))
+            return new ResponseEntity(new MensajeDTO("ese nombre ya existe"), HttpStatus.BAD_REQUEST);
+       articuloManufacturadoService.save(manuf);
+       return new ResponseEntity(new MensajeDTO("producto guardado"), HttpStatus.CREATED);
+    }
+	
+	
+	@Override
+	@PutMapping("/{id}")
+	//@PreAuthorize("hasRole('ADMIN')")
+	public ResponseEntity update(@RequestBody ArticuloManufacturadoDTO t, @PathVariable int id) {
+		return ResponseEntity.status(201).body(articuloManufacturadoService.update(t, id));
 	}
 	
-	
 	/*@Override
-	@PutMapping("/{id}")
-	@PreAuthorize("hasRole('ADMIN')")
-	public ResponseEntity updateEntity(@RequestBody ArticuloManufacturadoDTO t, @PathVariable int id) {
-		return ResponseEntity.status(201).body(articuloManufacturadoService.update(t, id));
-	}*/
-	
-	
-	@PutMapping("/{id}")
-    @PreAuthorize("hasRole('ADMIN')")
-	public ResponseEntity<?> updateEntity(@RequestBody ArticuloManufacturadoDTO manuf, @PathVariable("id") int id){
+	@PutMapping("/actualizar/{id}")
+    //@PreAuthorize("hasRole('ADMIN')")
+	public ResponseEntity<?> update(@RequestBody ArticuloManufacturadoDTO manuf, @PathVariable("id") int id){
         if(!articuloManufacturadoService.existePorId(id))
-            return new ResponseEntity(new MensajeDTO("no existe ese producto"), HttpStatus.NOT_FOUND);
+            return new ResponseEntity(new MensajeDTO("no existe ese articulo manufacturado"), HttpStatus.NOT_FOUND);
         if(StringUtils.isBlank(manuf.getNombre_articuloM()))
             return new ResponseEntity(new MensajeDTO("el nombre es obligatorio"), HttpStatus.BAD_REQUEST);
         if((Float)manuf.getPrecioM() == null || manuf.getPrecioM()==0)
@@ -92,23 +122,20 @@ public class ArticuloManufacturadoController implements ObjectController<Articul
         prodUpdate.setNombre_articuloM(manuf.getNombre_articuloM());
         prodUpdate.setPrecioM(manuf.getPrecioM());
         articuloManufacturadoService.save(prodUpdate);
-        return new ResponseEntity(new MensajeDTO("producto actualizado"), HttpStatus.CREATED);
-    }
+        return new ResponseEntity(new MensajeDTO("articulo manufacturado actualizado"), HttpStatus.CREATED);
+    }*/
 	
 	
-	
-
 	@Override
-	@DeleteMapping("/{id}")
-	@PreAuthorize("hasRole('ADMIN')")
-	public ResponseEntity deleteEntity(@PathVariable int id) {
-		boolean det = articuloManufacturadoService.delete(id);
-
-		if (det) {
-			return ResponseEntity.status(204).body("'Message' : 'Successful Delete'");
-		} else {
-			return ResponseEntity.status(400).body("'Message' . 'Unsuccessful delete'");
-		}
+	@DeleteMapping("/borrar/{id}")
+	//@PreAuthorize("hasRole('ADMIN')")
+	public ResponseEntity<?> delete(@PathVariable int id) {
+		
+		if(!articuloManufacturadoService.existePorId(id))
+            return new ResponseEntity(new MensajeDTO("no existe ese articulo"), HttpStatus.NOT_FOUND);
+        articuloManufacturadoService.delete(id);
+        return new ResponseEntity(new MensajeDTO("articulo eliminado"), HttpStatus.OK);
+        
 	}
 
 }
